requirements.txt
Flask
requests
dnspython
shodan
